GroundTile1 with texture (texture not finished)
I'm working on making this a tiling texture so of it is placed next to another tile you will not see a seam line. I will have that finished today. This is just in case you need the file soon. I exported using the same scale that you gave me the files at.